Update quarantine-manager: single command with bulk select
Consolidate the Check Quarantine Status and Remove Quarantine commands into
one Manage Quarantine command, and add uninstaller-style multi-select so you
can pick exactly which quarantined files to clear.

Implements the follow-up request in #25892.
